Look at the extended description...
Fixed: Parameters in the functions, since the dictionary is a global variable and we don't need to pass it
along with some other small fixes.
Functions:
wordInsetion was made to insert the words in the array. Idk if anything is inserted right but i print the words for testing and they seem alright, still not 100% sure (no segmentations).
Also, I tried to do the searching function which im sure it doesn't work 100% right, since when i try to search a word from the dictionary (a word that i see from the printing earlier), sometimes it prints the message that the word is not in the dictionary.
And also, the index is the hash so i think is uneccessary, maybe i need some explenation about that.